fixed warnings
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@823 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/ui/CloseButton.java
@@ -52,7 +52,7 @@ public class CloseButton extends GradiantButton
         super.paint(g);
         
         FontMetrics fm = this.getFontMetrics(getFont());
-        int w = (int)((double)fm.getAscent() * 0.8);
+        int w = (int)(fm.getAscent() * 0.8);
         Rectangle r = new Rectangle((getWidth() - w) / 2, (getHeight() - w) / 2, w, w);
         
         if (pressed)
